Modified favorite functionality

+++ careconnect/core/views.py
@@ -6,6 +6,7 @@ from .models import Favorite, Location, Review
 from rest_framework.response import Response
 import os
 from django.conf import settings
+from healthcare.models import Doctor, Hospital
 
 # Create your views here.
 class FavoriteViewSet(viewsets.ModelViewSet):
@@ -13,15 +14,47 @@ class FavoriteViewSet(viewsets.ModelViewSet):
     permission_classes = [AllowAny]
     queryset = Favorite.objects.all()
 
-    def destroy(self, request, *args, **kwargs):
-        try:
-            favorite = self.get_object()
+    def create(self, request, *args, **kwargs):
+        user_id = request.data.get('user_id')
+        entity_id = request.data.get('entity_id')
+        entity_type = request.data.get('entity_type')
+
+        # Check if the favorite already exists
+        favorite = Favorite.objects.filter(user_id=user_id, entity_id=entity_id, entity_type=entity_type).first()
+        if favorite:
+            # Toggle is_favorite field for the corresponding entity
+            if entity_type == 1:  # Doctor
+                doctor = Doctor.objects.filter(id=entity_id).first()
+                if doctor:
+                    doctor.is_favorite = not doctor.is_favorite
+                    doctor.save()
+            elif entity_type == 2:  # Hospital
+                hospital = Hospital.objects.filter(id=entity_id).first()
+                if hospital:
+                    hospital.is_favorite = not hospital.is_favorite
+                    hospital.save()
+            
             favorite.delete()
-            return Response({"message":"Favorite item removed sucessfully"}, status=status.HTTP_200_OK)
-        except Exception:
-            return Response({"error":"Favorite item not found"}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"message": "Favorite item removed successfully"}, status=status.HTTP_200_OK)
         
-        
+        # Create new favorite and set is_favorite field
+        favorite = Favorite(user_id=user_id, entity_id=entity_id, entity_type=entity_type)
+        favorite.save()
+
+        if entity_type == 1:  # Doctor
+            doctor = Doctor.objects.filter(id=entity_id).first()
+            if doctor:
+                doctor.is_favorite = True
+                doctor.save()
+        elif entity_type == 2:  # Hospital
+            hospital = Hospital.objects.filter(id=entity_id).first()
+            if hospital:
+                hospital.is_favorite = True
+                hospital.save()
+
+        return Response({"message": "Favorite item added successfully"}, status=status.HTTP_201_CREATED)
+
+          
 class LocationViewSet(viewsets.ModelViewSet):
     serializer_class = LocationSerializer
     permission_classes = [AllowAny]
